Radiation Calculation Rewrite
Cleaned it up ig
Made all methods use values pulled from QuantumConfig
Calculates all registered radiation types instead of hard coding them

+++ src/main/java/net/timeworndevs/quantum/util/ArmorInsulator.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 
 
-public record ArmorInsulator (ArrayList<Item> armorItems, ArrayList<Float> multipliers, HashMap<RadiationType, Float> radiations) {
+public record ArmorInsulator (ArrayList<Item> armorItems, ArrayList<Float> multipliers, HashMap<RadiationType, Integer> radiations) {
 
     public boolean containsItem(Item item) {
         return armorItems.contains(item);
